+++ src/controllers/discount.controller.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const DiscountService = require('../services/discount.service')
-const { SuccessResponse, OK, CREATED } = require("../core/success.response")
+const { SuccessResponse, OK, CREATED } = require("../helpers/success.response")
 
 class DiscountController {
