+++ src/controllers/product.controller.js
@@ -2,7 +2,7 @@
 
 const ProductService = require("../services/product.service")
 const ProductServiceV2 = require("../services/product.service.xxx")
-const { SuccessResponse, OK, CREATED } = require("../core/success.response")
+const { SuccessResponse, OK, CREATED } = require("../helpers/success.response")
 
 class ProductController {
