--- src/loggers/logger.discord.js
@@ -1,3 +1,0 @@
-'use strict'
-
-const a = 'nothing'
