--- src/postman/access.post.http
@@ -1,12 +1,0 @@
-@url_dev = http://localhost:3056/v1/api
-
-### signup
-POST {{url_dev}}/shop/signup
-Content-Type: application/json
-x-api-key: abde5a6bcf43aef84d8d0ccd1380b0a539a53fb77a89321c1aac7922a04e4c7d445057e750eee388f2bbe99a1e8334af283e4e2efdeb81601cc34c6ad87b9def
-
-{
-    "name": "Shop HieuDH",
-    "email": "danghieuyt1@gmail.com",
-    "password": "ctmu20hnsa"
-}
