+++ src/routes/access/index.js
@@ -3,8 +3,8 @@
 const express = require('express')
 const router = express.Router()
 const accessController = require('../../controllers/access.controller')
-const { asyncHandler } = require('../../auth/checkAuth')
-const { authentication, authenticationV2 } = require('../../auth/authUtils')
+const { asyncHandler } = require('../../middlewares/checkAuth')
+const { authentication, authenticationV2 } = require('../../middlewares/authUtils')
 
 // sign up
 router.post('/shop/signup', asyncHandler( accessController.signUp ))
