+++ src/routes/cart/index.js
@@ -3,8 +3,8 @@
 const express = require('express')
 const router = express.Router()
 const cartController = require('../../controllers/cart.controller')
-const { asyncHandler } = require('../../auth/checkAuth')
-const { authentication, authenticationV2 } = require('../../auth/authUtils')
+const { asyncHandler } = require('../../middlewares/checkAuth')
+const { authentication, authenticationV2 } = require('../../middlewares/authUtils')
 
 // authentication
 router.use(authenticationV2)
