+++ src/routes/checkout/index.js
@@ -2,8 +2,8 @@
 
 const express = require('express')
 const router = express.Router()
-const { asyncHandler } = require('../../auth/checkAuth')
-const { authentication, authenticationV2 } = require('../../auth/authUtils')
+const { asyncHandler } = require('../../middlewares/checkAuth')
+const { authentication, authenticationV2 } = require('../../middlewares/authUtils')
 const checkoutController = require('../../controllers/checkout.controller')
 
 // authentication
