+++ src/routes/discount/index.js
@@ -3,8 +3,8 @@
 const express = require('express')
 const router = express.Router()
 const discountController = require('../../controllers/discount.controller')
-const { asyncHandler } = require('../../auth/checkAuth')
-const { authentication, authenticationV2 } = require('../../auth/authUtils')
+const { asyncHandler } = require('../../middlewares/checkAuth')
+const { authentication, authenticationV2 } = require('../../middlewares/authUtils')
 
 // get amount a discount
 router.post('/amount', asyncHandler( discountController.getDiscountAmount ))
